Split dependencies such that they can be installed optionally for APIs (#379)
* Split dependencies such that they can be installed optionally for APIs

* Add pip upgrade step to github workflows

* Add llvmlite to deps

* Move plotting functions to utils
