+++ .github/workflows/docs_deploy.yml
@@ -26,7 +26,7 @@ jobs:
           sudo apt-get install libcurl4-openssl-dev libssl-dev pandoc
       - uses: actions/checkout@v3
       - name: Install dependencies, build docs and coverage report
-        run: pipx install poetry
+        run: python3 -m pip install --upgrade pip && python3 -m pip install poetry
       - uses: actions/setup-python@v4.3.0
         with:
           python-version: 3.9
