+++ cyclops/query/post_process/util.py
@@ -1,39 +1,10 @@
 """Post-processing functions applied to queried data (Pandas DataFrames)."""
 
-from typing import Union
-
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-from matplotlib.cm import get_cmap
-from matplotlib.pyplot import figure
 from pandas import Timestamp
 
 from cyclops.process.column_names import CARE_UNIT
-from cyclops.process.util import has_columns
-
-
-def to_timestamp(data: Union[pd.Series, np.ndarray]) -> pd.Series:
-    """Convert a Pandas series or NumPy array to a datetime/timestamp type.
-
-    Parameters
-    ----------
-    data: pandas.Series or numpy.ndarray
-        Data to be converted.
-
-    Returns
-    -------
-    pandas.Series
-        The converted data.
-
-    """
-    if isinstance(data, pd.Series):
-        return pd.to_datetime(data)
-
-    if isinstance(data, np.ndarray):
-        return pd.to_datetime(pd.Series(data))
-
-    raise ValueError(f"Type of data argument ({type(data)}) not supported.")
 
 
 def event_time_between(
@@ -78,50 +49,6 @@ def event_time_between(
         discharge_cond = event < discharge
 
     return admit_cond & discharge_cond
-
-
-def plot_admit_discharge(
-    data: pd.DataFrame, description: str = "description", figsize: tuple = (10, 4)
-) -> None:
-    """Plot a series of admit discharge times given a description.
-
-    Parameters
-    ----------
-    data: pandas.DataFrame
-        DataFrame with 'admit', 'discharge', and description columns.
-        The admit and discharge columns must be convertable to Timestamps.
-
-    """
-    data = data.copy()
-    has_columns(data, ["admit", "discharge", description], raise_error=True)
-
-    figure(figsize=figsize, dpi=80)
-    colors = get_cmap("Accent").colors
-
-    data["admit_int"] = to_timestamp(data["admit"]).astype(int)
-    data["discharge_int"] = to_timestamp(data["discharge"]).astype(int)
-
-    desc_dict = {}
-    for val, key in enumerate(data[description].unique()):
-        desc_dict[key] = val
-
-    data["admit_int"] = data["admit"].astype(int)
-    data["discharge_int"] = data["discharge"].astype(int)
-
-    plotted = []
-
-    def plot_timerange(admit, discharge, desc):
-        ind = desc_dict[desc]
-        if desc in plotted:
-            plt.plot([admit, discharge], [ind, ind], color=colors[ind])
-        else:
-            plt.plot([admit, discharge], [ind, ind], color=colors[ind], label=desc)
-            plotted.append(desc)
-
-    for _, row in data.iterrows():
-        plot_timerange(row["admit_int"], row["discharge_int"], row[description])
-
-    plt.legend()
 
 
 def process_care_unit_changepoints(
